+++ CHANGE_LOG.md
@@ -1,3 +1,8 @@
+# v0.1.0
+- Rename package name from `com.example.mywords` to `com.mywords.android`.
+- Initial version, ready to be launched on Google Play Store. Toady is a good day, 2024-10-19.
+- Add GitHub link to the `About` dialog.
+
 # v3.0.1
 - **Real-Time synchronization (Beta)**:  Add a source parameter to the Go API functions to discriminate between web and client versions. So it can keep real-time synchronization between the two versions when one version updated the data.
 - Due to the time difference between the underlying library and flutter, todayInt in flutter may be inaccurate, use the time of the underlying library, pass 1 to represent today